Add bootstrapping logic for ignition
